+++ R/tar_progress_summary.R
@@ -1,20 +1,15 @@
-#' @title Tabulate the progress of dynamic branches.
+#' @title Summarize target progress.
 #' @export
-#' @description Read a project's target progress data for the most recent
-#'   run of the pipeline and display the tabulated status
-#'   of dynamic branches. Only the most recent record is shown.
-#' @return A data frame with one row per target per progress status
-#'   and the following columns.
-#'   * `name`: name of the pattern.
-#'   * `progress`: progress status: `"started"`, `"built"`, `"cancelled"`,
-#'     or `"errored"`.
-#'   * `branches`: number of branches in the progress category.
-#'   * `total`: total number of branches planned for the whole pattern.
-#'     Values within the same pattern should all be equal.
-#' @param names Optional, names of the targets. If supplied, `tar_progress()`
-#'   only returns progress information on these targets.
-#'   You can supply symbols, a character vector,
-#'   or `tidyselect` helpers like [starts_with()].
+#' @description Summarize the progress of a run of the pipeline.
+#' @return A data frame with one row and the following optional columns that can be
+#'   selected with `fields`. (`time` is omitted by default.)
+#'   * `started`: number of targets that started and did not (yet) finish.
+#'   * `built`: number of targets that completed without error or cancellation.
+#'   * `errored`: number of targets that threw an error.
+#'   * `canceled`: number of canceled targets (see [tar_cancel()]).
+#'   * `since`: how long ago progress last changed (`Sys.time() - time`).
+#'   * `time`: the time when the progress last changed
+#'     (modification timestamp of the `_targets/meta/progress` file).
 #' @param fields Optional, names of progress data columns to read.
 #'   Set to `NULL` to read all fields.
 #' @examples
@@ -24,27 +19,32 @@
 #'   list(
 #'     tar_target(x, seq_len(2)),
 #'     tar_target(y, x, pattern = map(x)),
-#'     tar_target(z, stopifnot(y < 1.5), pattern = map(y))
+#'     tar_target(z, stopifnot(y < 1.5), pattern = map(y), error = "continue")
 #'   )
 #' }, ask = FALSE)
 #' try(tar_make())
-#' tar_progress_branches()
+#' tar_progress_summary()
 #' })
 #' }
-tar_progress_branches <- function(names = NULL, fields = NULL) {
+tar_progress_summary <- function(
+  fields = c("started", "built", "errored", "canceled", "since")
+) {
   assert_store()
   assert_path(path_progress())
-  out <- tibble::as_tibble(progress_init()$database$read_condensed_data())
-  out <- tar_progress_branches_summary(out)
-  names_quosure <- rlang::enquo(names)
+  time <- file.mtime(path_progress())
+  progress <- tibble::as_tibble(progress_init()$database$read_condensed_data())
+  progress <- progress[progress$type != "pattern",, drop = FALSE] # nolint
+  out <- tibble::tibble(
+    started = sum(progress$progress == "started"),
+    built = sum(progress$progress == "built"),
+    errored = sum(progress$progress == "errored"),
+    canceled = sum(progress$progress == "canceled"),
+    since = format_seconds(difftime(Sys.time(), time, units = "secs")),
+    time = time_stamp(time)
+  )
   fields_quosure <- rlang::enquo(fields)
-  names <- eval_tidyselect(names_quosure, out$name)
   fields <- eval_tidyselect(fields_quosure, colnames(out)) %|||% colnames(out)
-  if (!is.null(names)) {
-    out <- out[match(names, out$name),, drop = FALSE] # nolint
-  }
-  out <- out[, base::union("name", fields), drop = FALSE]
-  out[order(out$name),, drop = FALSE] # nolint
+  out[, fields, drop = FALSE]
 }
 
 tar_progress_branches_summary <- function(progress) {
@@ -88,9 +88,9 @@ tar_progress_branches_gt <- function() {
     locations = gt::cells_column_labels(everything())
   )
   colors <- data_frame(
-    progress = c("started", "built", "canceled", "errored"),
-    fill = c("#DC863B", "#E1BD6D", "#FAD510", "#C93312"),
-    color = c("black", "black", "black", "white")
+    progress = c("started", "built", "errored", "canceled"),
+    fill = c("#DC863B", "#E1BD6D", "#C93312", "#FAD510"),
+    color = c("black", "black", "white", "black")
   )
   for (index in seq_len(nrow(colors))) {
     out <- gt::tab_style(