+++ R/tar_traceback.R
@@ -4,9 +4,9 @@
 #'   `tar_traceback()` returns its traceback. The workspace file must
 #'   exist. For more information, see [tar_workspace()].
 #' @return Character vector, the traceback of a failed target
-#'   if it exists
+#'   if it exists.
 #' @param name Symbol, name of the target whose workspace to read.
-#' @param envir Environment in which to put the objects.
+#' @param envir Deprecated in `targets` > 0.3.1 (2021-03-28).
 #' @param packages Logical, whether to load the required packages
 #'   of the target.
 #' @param source Logical, whether to run `_targets.R` to load user-defined
@@ -25,19 +25,34 @@
 #'   )
 #' }, ask = FALSE)
 #' try(tar_make())
-#' tar_traceback(y)
+#' tail(tar_traceback(y))
 #' })
 #' }
 tar_traceback <- function(
   name,
-  envir = parent.frame(),
-  packages = TRUE,
-  source = TRUE
+  envir = NULL,
+  packages = NULL,
+  source = NULL
 ) {
-  force(envir)
+  if (!is.null(envir) || !is.null(packages) || !is.null(source)) {
+    warn_deprecate(
+      "The envir, packages, and source arguments of tar_traceback() ",
+      "are deprectaed in targets > 0.3.1 (2021-03-28)."
+    )
+  }
   name <- deparse_language(substitute(name))
   assert_chr(name)
   assert_scalar(name)
   workspace <- workspace_read(name)
-  workspace$target$metrics$traceback
+  out <- workspace$target$metrics$traceback
+  if (is.null(out)) {
+    return(character(0))
+  }
+  n <- length(out)
+  min <- max(which(grepl("^build_eval_fce17be7", out))) %||% 1 %||NA% 1
+  max <- max(which(grepl("^\\.handleSimpleError", out))) %||% n %||NA% n
+  if (length(min) == 1L && length(max) == 1L && (max - min) >= 2L) {
+    out <- out[seq(min + 1, max - 1)]
+  }
+  out
 }